v1.0/requirements.txt
Authlib==1.0.1
Flask==2.2.2
Flask_HTTPAuth==4.7.0
Flask_Login==0.6.2
Flask_Mail==0.9.1
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.1
fpdf==1.7.2
gevent==22.8.0
google_auth_oauthlib==0.5.2
itsdangerous==2.1.2
numpy==1.23.3
Pillow==9.2.0
pip==22.2.2
protobuf==4.21.7
requests==2.28.1
SQLAlchemy==1.4.40
